refactor: extract validator logic to custom hook
`TodoInput`컴포넌트의 유효성 검사 로직을 `useValidator`라는 커스텀 훅으로 분리함.
